Main Project Part 11
adding a trail to our game using a trail renderer which will eventually turn into a slice, applying materials on it also.

Also trying to fix the camera here using a built in unity function that checks to see if the mouse is being held and transfer those cordinates to the game using screentoworldpoint.

Getting rid of userInputHandler as I will add to this at a later point for optimising touch input.